Chore (./): Formating code

+++ src/UniswapV2Factory.sol
@@ -5,25 +5,24 @@ import {IUniswapV2Factory} from "@main/interfaces/IUniswapV2Factory.sol";
 import {UniswapV2PairVault} from "@main/UniswapV2PairVault.sol";
 
 contract UniswapV2Factory is IUniswapV2Factory {
-
     uint8 public flashLoanFee;
 
     mapping(address => mapping(address => address)) public getPair;
     address[] public allPairs;
 
-    constructor(uint8 _flashLoanFee ) {
+    constructor(uint8 _flashLoanFee) {
         flashLoanFee = _flashLoanFee;
     }
 
-    function allPairsLength() external view returns (uint) {
+    function allPairsLength() external view returns (uint256) {
         return allPairs.length;
     }
 
     function createPair(address tokenA, address tokenB) external returns (address pair) {
-        require(tokenA != tokenB, 'UniswapV2: IDENTICAL_ADDRESSES');
+        require(tokenA != tokenB, "UniswapV2: IDENTICAL_ADDRESSES");
         (address token0, address token1) = tokenA < tokenB ? (tokenA, tokenB) : (tokenB, tokenA);
-        require(token0 != address(0), 'UniswapV2: ZERO_ADDRESS');
-        require(getPair[token0][token1] == address(0), 'UniswapV2: PAIR_EXISTS'); // single check is sufficient
+        require(token0 != address(0), "UniswapV2: ZERO_ADDRESS");
+        require(getPair[token0][token1] == address(0), "UniswapV2: PAIR_EXISTS"); // single check is sufficient
         bytes memory bytecode = type(UniswapV2PairVault).creationCode;
         bytes32 salt = keccak256(abi.encodePacked(token0, token1));
         assembly {
@@ -35,6 +34,4 @@ contract UniswapV2Factory is IUniswapV2Factory {
         allPairs.push(pair);
         emit PairCreated(token0, token1, pair, allPairs.length);
     }
-
-
 }
